requirements.txt
asgiref==3.6.0
beautifulsoup4==4.11.2
certifi==2022.12.7
charset-normalizer==3.1.0
Django==4.1.5
django-environ==0.10.0
djangorestframework==3.14.0
idna==3.4
psycopg2==2.9.5
pytz==2022.7.1
requests==2.28.2
soupsieve==2.4
sqlparse==0.4.3
tzdata==2022.7
urllib3==1.26.15
